perf(core): migrate to apollo 3 for more graphql stability and perfs
