--- apps/client/src/app/apollo-interceptor.ts
@@ -1,30 +1,0 @@
-import { Injectable } from '@angular/core';
-import { HttpHandler, HttpInterceptor, HttpRequest } from '@angular/common/http';
-import { AuthFacade } from './+state/auth.facade';
-import { filter, first, switchMap } from 'rxjs/operators';
-
-@Injectable()
-export class ApolloInterceptor implements HttpInterceptor {
-
-  constructor(private authFacade: AuthFacade) {
-  }
-
-  intercept(req: HttpRequest<any>, next: HttpHandler) {
-    if (req.url.indexOf('gubal.') > -1) {
-      return this.authFacade.idToken$
-        .pipe(
-          filter((token: any) => token.claims['https://hasura.io/jwt/claims'] !== undefined),
-          first(),
-          switchMap(idToken => {
-            const clone = req.clone({
-              setHeaders: {
-                Authorization: `Bearer ${idToken.token}`
-              }
-            });
-            return next.handle(clone);
-          }));
-    }
-    return next.handle(req);
-  }
-}
-
