--- apps/client/src/app/graphql.module.ts
@@ -1,9 +1,0 @@
-import { NgModule } from '@angular/core';
-import { ApolloModule } from 'apollo-angular';
-import { HttpLinkModule } from 'apollo-angular-link-http';
-
-@NgModule({
-  exports: [ApolloModule, HttpLinkModule]
-})
-export class GraphQLModule {
-}
