+++ apps/client/src/app/pages/fish-train/fish-train-routing.module.ts
@@ -1,7 +1,6 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { FishTrainComponent } from './fish-train/fish-train.component';
-import { ApolloClientResolver } from '../../core/apollo-client.resolver';
 import { MaintenanceGuard } from '../maintenance/maintenance.guard';
 import { VersionLockGuard } from '../version-lock/version-lock.guard';
 
@@ -9,10 +8,7 @@ const routes: Routes = [
   {
     path: ':id',
     component: FishTrainComponent,
-    canActivate: [MaintenanceGuard, VersionLockGuard],
-    resolve: {
-      client: ApolloClientResolver
-    }
+    canActivate: [MaintenanceGuard, VersionLockGuard]
   }
 ];
