+++ package.json
@@ -165,6 +165,7 @@
     "@angular/platform-browser-dynamic": "14.2.6",
     "@angular/platform-server": "14.2.6",
     "@angular/router": "14.2.6",
+    "@apollo/client": "^3.7.9",
     "@ffxiv-teamcraft/simulator": "^3.1.6",
     "@ffxiv-teamcraft/simulator-cn": "npm:@ffxiv-teamcraft/simulator@3.0.14",
     "@ffxiv-teamcraft/simulator-kr": "npm:@ffxiv-teamcraft/simulator@3.0.14",
@@ -218,12 +219,7 @@
     "@xivapi/angular-client": "2.0.1",
     "algoliasearch": "^4.14.2",
     "angular2-chartjs": "^0.5.1",
-    "apollo-angular": "^1.8.0",
-    "apollo-angular-link-http": "^1.9.0",
-    "apollo-cache-inmemory": "^1.6.0",
-    "apollo-client": "^2.6.0",
-    "apollo-link": "^1.2.11",
-    "apollo-link-ws": "^1.0.20",
+    "apollo-angular": "^4.2.1",
     "axios": "^1.2.6",
     "bufferutil": "^4.0.5",
     "caniuse-lite": "^1.0.30001419",
